Issue006 : added UI package, move UI classes to UI package, renaming some classes and files
