--- app/src/main/java/org/edb/main/PluginConfigUI.java
@@ -1,4 +1,0 @@
-package org.edb.main;
-
-public class PluginConfigUI {
-}
