+++ app/src/main/java/org/edb/main/UI/BootApp.java
@@ -1,18 +1,13 @@
-package org.edb.main;
+package org.edb.main.UI;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.SplitPane;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.BorderPane;
-import javafx.scene.layout.Pane;
 import javafx.stage.Stage;
+import org.edb.main.Authorization;
 
-import java.io.IOException;
-
-public class MainUI extends Application {
+public class BootApp extends Application {
 
 
 
@@ -26,7 +21,7 @@ public class MainUI extends Application {
 
         if(token == ""){
             //token이 없으면
-            Parent root = FXMLLoader.load(getClass().getResource("../../../fxml/sceneMain.fxml"));
+            Parent root = FXMLLoader.load(getClass().getResource("/fxml/mainUI.fxml"));
 
             Scene scene = new Scene(root);
 
@@ -35,7 +30,7 @@ public class MainUI extends Application {
         }else{
 
             //token이 있으면
-            Parent root = FXMLLoader.load(getClass().getResource("../../../fxml/sceneMainAfterLogin.fxml"));
+            Parent root = FXMLLoader.load(getClass().getResource("/fxml/mainUIAfterLogin.fxml"));
 
             Scene scene = new Scene(root);
