+++ app/src/main/java/org/edb/main/UI/Controller.java
@@ -1,4 +1,4 @@
-package org.edb.main;
+package org.edb.main.UI;
 
 public class Controller {
 }
