+++ app/src/main/java/org/edb/main/UI/PluginConfigUIController.java
@@ -1,0 +1,4 @@
+package org.edb.main.UI;
+
+public class PluginConfigUIController {
+}
