--- app/src/main/java/org/edb/main/externalServiceContainer.java
@@ -1,4 +1,0 @@
-package org.edb.main;
-
-public class externalServiceContainer {
-}
