+++ app/src/main/resources/fxml/mainUIAfterLogin.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.control.Button?>
 
 <?import javafx.scene.layout.BorderPane?>
-<BorderPane fx:controller="org.edb.main.mainUserinfoSubController" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171">
+<BorderPane fx:controller="org.edb.main.UI.mainUserinfoSubController" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171">
 
     <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="160.0" prefWidth="100.0">
         <children>
