+++ app/src/main/resources/fxml/sample.fxml
@@ -3,6 +3,6 @@
 
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
-<GridPane fx:controller="org.edb.main.Controller"
+<GridPane fx:controller="org.edb.main.UI.Controller"
           xmlns:fx="http://javafx.com/fxml" alignment="center" hgap="10" vgap="10">
 </GridPane>
